+++ Collection.php
@@ -827,6 +827,7 @@ class Collection extends Object
     public function buildCondition($condition)
     {
         static $builders = [
+            'NOT' => 'buildNotCondition',
             'AND' => 'buildAndCondition',
             'OR' => 'buildOrCondition',
             'BETWEEN' => 'buildBetweenCondition',
@@ -887,6 +888,34 @@ class Collection extends Object
                     $result[$name] = $value;
                 }
             }
+        }
+
+        return $result;
+    }
+
+    /**
+     * Composes `NOT` condition.
+     * @param string $operator the operator to use for connecting the given operands
+     * @param array $operands the Mongo conditions to connect.
+     * @return array the generated Mongo condition.
+     * @throws InvalidParamException if wrong number of operands have been given.
+     */
+    public function buildNotCondition($operator, $operands)
+    {
+        if (!isset($operands[0], $operands[1])) {
+            throw new InvalidParamException("Operator '$operator' requires two operands.");
+        }
+
+        list($name, $value) = $operands;
+
+        $result = [];
+        if (is_array($value)) {
+            $result[$name] = ['$not' => $this->buildCondition($value)];
+        } else {
+            if ($name == '_id') {
+                $value = $this->ensureMongoId($value);
+            }
+            $result[$name] = ['$ne' => $value];
         }
 
         return $result;